+++ LHCb/Event/L0Event/src/L0DUElementaryData.cpp
@@ -1,4 +1,4 @@
-// $Id: L0DUElementaryData.cpp,v 1.6 2008-05-29 14:10:07 odescham Exp $
+// $Id: L0DUElementaryData.cpp,v 1.7 2008-09-26 13:30:22 odescham Exp $
 // Include files 
 #include <utility>
 #include <string>
@@ -13,8 +13,9 @@
 // 09/02/2006 : Olivier Deschamps
 //-----------------------------------------------------------------------------
 
-long LHCb::L0DUElementaryData::operation(const std::string& v_operator, const std::vector<long>& v_operands){
-  long digit = 0;
+unsigned int LHCb::L0DUElementaryData::operation(const std::string& v_operator, const std::vector<unsigned int>& 
+v_operands){
+  unsigned int digit = 0;
   // 1 dimensional operator (Identity)
   if(1 == v_operands.size() ){
     if( "Id" == v_operator ){digit = v_operands[0] ;}
